+++ packages/ui-service-inquirer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/ui-service-inquirer
 
+## 0.2.51
+
+### Patch Changes
+
+-   Updated dependencies [1357412]
+    -   @sap-ux/odata-service-inquirer@2.11.6
+
 ## 0.2.50
 
 ### Patch Changes